+++ android/TodokTodok/app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:theme="@style/Theme.TodokTodok"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.view.searchbook.SearchBooksActivity"
+            android:name=".presentation.view.searchbooks.SearchBooksActivity"
             android:exported="false" />
         <activity
             android:name=".presentation.view.MainActivity"
