+++ android/TodokTodok/app/src/main/java/com/example/todoktodok/presentation/view/searchbooks/OnSelectBookListener.kt
@@ -1,4 +1,4 @@
-package com.example.todoktodok.presentation.view.searchbook
+package com.example.todoktodok.presentation.view.searchbooks
 
 fun interface OnSelectBookListener {
     fun select(id: Long)
