+++ android/TodokTodok/app/src/main/java/com/example/todoktodok/presentation/view/searchbooks/SearchBooksAdapter.kt
@@ -1,9 +1,9 @@
-package com.example.todoktodok.presentation.view.searchbook
+package com.example.todoktodok.presentation.view.searchbooks
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
 import com.example.domain.model.Book
-import com.example.todoktodok.presentation.view.searchbook.SearchBooksViewHolder.Companion.SearchBooksViewHolder
+import com.example.todoktodok.presentation.view.searchbooks.SearchBooksViewHolder.Companion.SearchBooksViewHolder
 
 class SearchBooksAdapter(
     private val onSelectBookListener: OnSelectBookListener,
