+++ android/TodokTodok/app/src/main/java/com/example/todoktodok/presentation/view/searchbooks/SearchBooksViewHolder.kt
@@ -1,4 +1,4 @@
-package com.example.todoktodok.presentation.view.searchbook
+package com.example.todoktodok.presentation.view.searchbooks
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
